requirements.txt
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
python-dotenv==0.17.1
PyYAML==5.4.1
uritemplate==3.0.1
